moving extensions into a dedicated directory
