+++ pyproject.toml
@@ -46,15 +46,16 @@ dependencies = [
     "tables==3.6.1; python_version<\"3.9\"",
     "tables>=3.7.0; python_version>\"3.8\"",
     "toml",
+    "qtconsole"
 ]
 
 [project.scripts]
-daq_logger = "pymodaq.daq_logger:main"
+daq_logger = "pymodaq.extensions.daq_logger:main"
 daq_move = "pymodaq.control_modules.daq_move:main"
-daq_scan = "pymodaq.daq_scan:main"
+daq_scan = "pymodaq.extensions.daq_scan:main"
 daq_viewer = "pymodaq.control_modules.daq_viewer:main"
 dashboard = "pymodaq.dashboard:main"
-h5browser = "pymodaq.h5browser:main"
+h5browser = "pymodaq.extensions.h5browser:main"
 parameter_example = "pymodaq.examples.parameter_ex:main"
 
 [project.urls]
