+++ src/pymodaq/extensions/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""
+Created the 25/10/2022
+
+@author: Sebastien Weber
+"""
