+++ src/pymodaq/extensions/h5browser.py
